requirements.txt
groq==0.11.0
requests==2.31.0
functions-framework==3.5.0
python-dotenv==1.0.0
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
httpx==0.27.0v